Move tubes selection state and logic from tubes-selection component to parent component generate-target-profile
